+++ app/src/main/java/com/monokoumacorporation/todoc/ui/create/button/OnProjectButton.java
@@ -1,5 +1,5 @@
 package com.monokoumacorporation.todoc.ui.create.button;
 
 public interface OnProjectButton {
-    void onProjectButtonClicked(int buttonId);
+    void onProjectButtonClicked(long buttonId);
 }
